requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements-txt -o requirements.txt
openai
azureopenai
instructor
pydantic
reportlab
streamlit
pypdf
pymupdf
frontend
tools
python-dotenv